back_end/requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
Flask-Cors
cloudinary
python-dotenv